requirements.txt
streamlit==1.28.1
requests==2.31.0
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3
pyserial==3.5
pywebview==4.4.1
pythonnet==3.0.3
flask==3.0.0
flask-cors==4.0.0